backend/pyproject.toml
[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[project]
name = "jpg_backend"
version = "0.3.0"
description = "Backend for JohnsGamePrice"
authors = [
    { name = "Jamie Hunt", email = "jamhun04@gmail.com" }
]
dependencies = [
    "requests",
    "python-dotenv",
    "fastapi",
    "uvicorn",
    "sqlalchemy (>=2.0.46,<3.0.0)",
    "alembic (>=1.18.1,<2.0.0)",
    "psycopg2 (>=2.9.11,<3.0.0)"
]
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"

[dependency-groups]
dev = [
    "pytest (>=9.0.2,<10.0.0)",
    "pytest-cov (>=7.0.0,<8.0.0)",
    "httpx (>=0.28.1,<0.29.0)",
    "black (>=25.12.0,<26.0.0)",
    "isort (>=7.0.0,<8.0.0)",
    "poethepoet (>=0.40.0,<0.41.0)"
]

[tool.black]
line-length = 88
target-version = ['py310']

[tool.isort]
profile = "black"

[tool.poe.tasks]
dev = "uvicorn jpg_backend.main:app --reload"
test = "pytest"
cov = "pytest --cov=jpg_backend --cov-report=term-missing"
format = ["_black", "_isort"]
_black = "black src tests"
_isort = "isort src tests"
check = ["_black_check", "_isort_check"]
_black_check = "black --check src tests"
_isort_check = "isort --check-only src tests"
